requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.2
django-redis==5.3.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
kombu==5.3.1
line-bot-sdk==2.4.2
multidict==6.0.4
openai==0.27.8

# psycopg2==2.9.6 # 編譯失敗，需要sudo apt-get install libpq-dev
psycopg2-binary==2.9.6

prompt-toolkit==3.0.38
python-dateutil==2.8.2
redis==4.5.5
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2